% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsdr_footnote_symbol.R
\name{atsdr_footnote_symbol}
\alias{atsdr_footnote_symbol}
\title{Footnote Symbols based on ATSDR Style Manual (2019)}
\usage{
atsdr_footnote_symbol(i = 0)
}
\arguments{
\item{i}{An index according to the symbol order of a footnote. Value of 0 gives the five unique footnote symbols in order of usage in the manual (useful if you just need the symbols themselves). If the index is a vector, a list of selected footnote symbols will be outputted. Default is 0.}
}
\value{
A character object.
}
\description{
Produces the proper footnote symbol based on a numeric index or numeric vector.
}
\examples{
atsdr_footnote_symbol(1)
atsdr_footnote_symbol(1:10)


}
