% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atsdr_recommended_cv}
\alias{atsdr_recommended_cv}
\title{PHAST Recommended CVs}
\format{
A 2,126 × 6 data frame:
\describe{
\item{contaminant}{Contaminant Name}
\item{casrn}{Contaminant CAS-Reference Number}
\item{cv}{Comparison Value}
\item{unit}{Comparison Value Unit of Measurement}
\item{cv_type}{Comparison Value Type}
\item{media}{Media/Matrix}
\item{phast_version}{PHAST version}
\item{database_rev}{Database Revision}
}
}
\usage{
atsdr_recommended_cv
}
\description{
PHAST Recommended CVs
}
\keyword{datasets}
