% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkjohn_correction.R
\name{barkjohn_correction}
\alias{barkjohn_correction}
\title{Barkjohn Correction Factor for PurpleAir monitors}
\usage{
barkjohn_correction(x, rh)
}
\arguments{
\item{x}{A value representing PurpleAir PM2.5 output}

\item{rh}{A value representing PurpleAir relative humidity}
}
\value{
A numeric class object
}
\description{
Applies the EPA standardized version of the Barkjohn correction factor for PurpleAir devices.
}
\examples{
barkjohn_correction(0.5, 30)

}
