% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bap_equivalents.R
\name{calculate_bap_equivalents}
\alias{calculate_bap_equivalents}
\title{Calculate Benzo(a)pyrene (BaP) Equivalents for Polycyclic Aromatic Hydrocarbons}
\usage{
calculate_bap_equivalents(obs, cen, casrn, sigfig = 4, message = TRUE)
}
\arguments{
\item{obs}{A numeric vector}

\item{cen}{A logical vector pertaining to censoring of obs. TRUE if obs is censored.}

\item{casrn}{A character vector representing the CAS-RN for each compound.}

\item{sigfig}{A numeric value for the number of significant figures for the outputs of the function. Default is 4.}

\item{message}{Do you want mesasge?}
}
\value{
A list class object.
\itemize{
\item \code{bap}: numeric value of the BaP Equivalent
\item \code{bap_lower}: numeric value of the lower bound BaP Equivalent when at least one compound has all nondetects.
\item \code{bap_upper}: numeric value of the upper bound BaP Equivalent when at least one compound has all nondetects.
\item \code{contaminant}: compound the BaP Equivalent should be compared to.
\item \code{casrn}: CAS-RN of the compound the BaP Equivalent should be compared to.
\item \code{notes}: character string of the underlying process that outputted the returned values.
\item \code{qcontrol}: character string of the potential errors related to the data in relation to the returned values.#'
}
}
\description{
Calculate BaP equivalents according to Guidance for Calculating Benzo(a)pyrene Equivalents for Cancer Evaluations of Polycyclic Aromatic Hydrocarbons (2022). Last updated 11/13/2024.
}
\details{
The function can be used in conjunction with dplyr \code{group_by()} and \code{summarise()} functions to output values not dependent on the CAS-RN inputted. Whether or not \code{casrn} argument has only BaP or a mix with other compounds, the function will only use values concerning BaP equivalent compounds, as long as the arguments are represented well.
}
\examples{
#example of individual inputs
results = rexp(n = 20, rate = 1)
nondetects <- results<0.5
congener <- rep(pef_congener$casrn, length.out = 20)

calculate_bap_equivalents(obs = results, cen = nondetects, casrn = congener)

#example of dplyr verbs with dataframe
library(tidyverse)

congener <- rep(pef_congener$casrn, each = 2, length.out = 20)
data.frame(results, nondetects, congener) |>
  summarise(bap = calculate_bap_equivalents(obs = results,
                                            cen = nondetects,
                                            casrn = congener)$bap)

}
