% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_epc.R
\name{calculate_epc}
\alias{calculate_epc}
\title{Calculate Exposure Point Concentrations (EPCs)}
\usage{
calculate_epc(
  obs = NULL,
  cen = NULL,
  sigfig = 4,
  testForNormal = TRUE,
  useDefaultSeed = TRUE
)
}
\arguments{
\item{obs}{A numeric vector}

\item{cen}{A logical vector pertaining to censoring of obs. TRUE if obs is censored.}

\item{sigfig}{A number for significant figures for the function. Default is 4.}

\item{testForNormal}{Logical. If you want to test for normal distribution of your obs.}

\item{useDefaultSeed}{Logical. TRUE uses my custom seed.}
}
\value{
A data.frame class object.
}
\description{
Calculate exposure point concentrations according to ATSDR Division of Community Health Investigations Exposure Point Concentration Guidance for Discrete Sampling (2023).
}
\examples{
set.seed(20240406)
results <- rexp(n = 15, rate = 1)
nondetects <- results<0.8
calculate_epc(obs = results, cen = nondetects)

}
