% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_teq.R
\name{calculate_teq}
\alias{calculate_teq}
\title{Calculate Toxic Equivalents (TEQ) for Dioxin and Dioxin-Like Compounds}
\usage{
calculate_teq(obs, cen, casrn, sigfig = 4, warnings = TRUE)
}
\arguments{
\item{obs}{A numeric vector}

\item{cen}{A logical vector pertaining to censoring of obs. TRUE if obs is censored.}

\item{casrn}{A character vector representing the CAS-RN for each compound.}

\item{sigfig}{A numeric value for the number of significant figures for the outputs of the function. Default is 4.}

\item{warnings}{Logical. Do you want warnings?}
}
\value{
A list class object.
\itemize{
\item \code{teq}: numeric value of the TEQ.
\item \code{teq_cen}: logical value of the censoring of the TEQ. TRUE for non-detect.
\item \code{rpd}: numeric value for the relative percent difference between the lower-bound TEQ estimate and the upper-bound TEQ estimate.
\item \code{rpd_flag}: flag for "RPD > 50\%".
\item \code{contaminant}: compound the TEQ should be compared to.
\item \code{casrn}: CAS-RN of the compound the TEQ should be compared to.
\item \code{notes}: character string of the underlying process that outputted the returned values.
\item \code{qcontrol}: character string of the potential errors related to the data in relation to the returned values.#'
}
}
\description{
Calculate TEQs according to ATSDR Toxic Equivalents Procedure for Dioxin and Dioxin-like Compounds Evaluation (2025). Last updated 07/25/2025.
}
\details{
The function can be used in conjunction with dplyr \code{group_by()} and \code{summarise()} functions to output values not dependent on the CAS-RN inputted. Whether or not \code{casrn} argument has exactly only dioxin and DLC or a mix with other compounds, the function will only use values concerning dioxin and DLC, as long as the arguments are represented well.
}
\examples{
results = rexp(n = 20, rate = 1)
nondetects <- results<0.5
congener <- rep(tef_congener$casrn, length.out = 20)

calculate_teq(obs = results, cen = nondetects, casrn = congener)

library(tidyverse)

congener <- rep(tef_congener$casrn, each = 2, length.out = 20)

data.frame(results, nondetects, congener, group = rep(c("A", "B"), 20)) |>
  group_by(group) |>
  summarise(teq = calculate_teq(obs = results, cen = nondetects, casrn = congener)$teq)

}
