% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_adjust.R
\name{ess_test}
\alias{ess_test}
\title{T-test using Effective Sample Size for Data with Autocorrelation}
\usage{
ess_test(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  method = c("basic", "hsdda"),
  warnings = TRUE
)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector. Optional for two sample tests.}

\item{alternative}{A character string for alternative hypothesis. Can be either "two.sided", "greater", or "less". Default is "two.sided".}

\item{conf.level}{A numeric value. For confidence level of the interval.}

\item{method}{A character string for method to calculate the factor equation for the effective samples sizes. Only "basic" and "hsdda" are available.}

\item{warnings}{A logical value. Change to remove warnings.}
}
\value{
A list class object (trying to figure out attributes to make my own class):
\itemize{
\item \code{statistic}: the t-statistic.
\item \code{parameter}: degrees of freedom.
\item \code{p.value}: p-value.
\item \code{conf.int}: confidence interval for the mean.
\item \code{estimate}: the mean.
\item \code{samplesize}: sample size.
\item \code{stderr}: the standard error of the mean.
\item \code{autocorr}: autocorrelation calculated through the auto-covariance and correlation function (ACF).
\item \code{ess.factor}: the calculated factor for effective sample size, used as a weighting term for the sample size.
\item \code{adj.statistic}: the t-statistic, adjusted using effective sample size.
\item \code{adj.parameter}: degrees of freedom, adjusted using effective sample size.
\item \code{adj.p.value}: p-value, adjusted using effective sample size.
\item \code{adj.conf.int}: confidence interval for the mean, adjusted using effective sample size.
}
}
\description{
Uses the t-test with effective sample size to calculate test statistic, p-value, and confidence intervals for autocorrelated data.
}
\examples{
x <- rnorm(10, 0, 5)
ess_test(x)
}
