% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_casrn.R
\name{fix_casrn}
\alias{fix_casrn}
\title{Fix Chemical Abstracts Service (CAS) Registry Number (RN)}
\usage{
fix_casrn(casrn, format = c("xx-xx-x", "xxxxx"))
}
\arguments{
\item{casrn}{A character vector representing CAS-RN.}

\item{format}{A character vector representing format for CAS-RN.}
}
\value{
A character string.
}
\description{
Formats CAS-RN. Able to check for whitespaces. Checks for CAS-RN validity.
}
\examples{
fix_casrn(c("39001-02-0", "74472-37-0"))
}
