% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jspr_install_packages.R
\name{jspr_install_packages}
\alias{jspr_install_packages}
\title{Install All Necessary Packages}
\usage{
jspr_install_packages(x = NULL)
}
\arguments{
\item{x}{A character vector with all names of packages to be added to the list}
}
\value{
An installation of an array of packages
}
\description{
Installs all packages required for many statistical analysis done at ATSDR.
}
\examples{
\dontrun{
jspr_install_packages()
}

}
