% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mission_area}
\alias{mission_area}
\title{Mission Area Categories and Keywords}
\format{
An 262 x 2 data frame:
\describe{
\item{category}{Mission Area Categories}
\item{use_for}{Description of Mission Area Categories}
\item{keyword}{Mission Area Keywords}
}
}
\usage{
mission_area
}
\description{
Mission Area Categories and Keywords
}
\keyword{datasets}
