% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pef_congener}
\alias{pef_congener}
\title{PEF List}
\format{
An 25 x 3 data frame:
\describe{
\item{casrn}{CAS-RN}
\item{pah}{PAH name}
\item{pef}{PEF value}
}
}
\usage{
pef_congener
}
\description{
PEF list
}
\keyword{datasets}
