% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommended_cv.R
\name{recommended_cv}
\alias{recommended_cv}
\title{ATSDR Recommended Screening CVs}
\usage{
recommended_cv(
  casrn,
  media = c("drinking_water", "shower", "soil", "air", "svi_soil_gas", "svi_groundwater"),
  unit = c("µg/m3", "ug/m3", "ppb", "ppm")
)
}
\arguments{
\item{casrn}{A character vector representing the CAS-RN for each compound.}

\item{media}{A character vector representing the media. should be one of the following: "drinking_water", "shower", "soil", "air", "svi_soil_gas", "svi_groundwater"}

\item{unit}{Optional character vector representing the units. Default NULL.}
}
\value{
A list class object.
\itemize{
\item \code{teq}: numeric value of the TEQ.
}
}
\description{
Produces recommended screening CVs given the compound and m. Last updated 04/21/2026.
}
\examples{
#example of individual inputs
recommended_cv(casrn = "7783-06-4", media = "air")$cv

#example of dplyr verbs with dataframe
library(tidyverse)

data <- data.frame(casrn = c("7783-06-4", "71-43-2", "7446-09-5"), media = "air")
data |> group_by(casrn, media) |> mutate(screen_cv = recommended_cv(casrn, media)$cv)

}
