% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_recommended_cv.R
\name{show_recommended_cv}
\alias{show_recommended_cv}
\title{ATSDR Recommended Screening CVs}
\usage{
show_recommended_cv(
  casrn,
  media = c("drinking_water", "shower", "soil", "air", "svi_soil_gas", "svi_groundwater"),
  unit = c("µg/m3", "ug/m3", "ppb", "ppm")
)
}
\arguments{
\item{casrn}{A character vector representing the CAS-RN for each compound.}

\item{media}{A character vector representing the media. should be one of the following: "drinking_water", "shower", "soil", "air", "svi_soil_gas", "svi_groundwater"}

\item{unit}{Optional character vector representing the units. Default NULL.}
}
\value{
A data.frame class object.
\itemize{
\item \code{casrn}: CAS reference number for the associated recommended CV.
\item \code{media}: Media associated to the CV.
\item \code{contaminant}: Contaminant name as in PHAST.
\item \code{cv}: numeric value of the recommended CV.
\item \code{unit}: Unit for CV.
\item \code{cv_type}: CV type of the recommended CV.
\item \code{phast_version}: PHAST version the data is pulled from.
\item \code{database_rev}: database revision for the PHAST database.
}
}
\description{
Produces recommended screening CVs given the compound and m. Last updated 04/21/2026.
}
\examples{
#example of individual inputs
show_recommended_cv(casrn = c("7783-06-4", "71-43-2", "67-56-1", "7664-41-7", "67-64-1"),
                    media = "air", unit = "ppb")

}
