% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tef_congener}
\alias{tef_congener}
\title{TEF List}
\format{
An 29 x 4 data frame:
\describe{
\item{casrn}{CAS-RN}
\item{congener_class}{Congener class}
\item{full_congener_name}{Dioxin and dioxin-like coupound chemical name}
\item{shorthand_congener_name}{Dioxin and dioxin-like coupound chemical shorthand name}
\item{tef}{TEF value}
}
}
\usage{
tef_congener
}
\description{
TEF list
}
\keyword{datasets}
