% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_marker.R
\name{add_metadata_marker}
\alias{add_metadata_marker}
\title{Insert Metadata Marker into README.md File}
\usage{
add_metadata_marker(
  contact_email,
  org = "ATSDR",
  version = NA,
  status = c("Maintained", "Deprecated"),
  keywords = NA,
  labor_hours = NA,
  contract = NA,
  exemption = NA,
  exemption_justification = NA
)
}
\arguments{
\item{contact_email}{Email for metadata or repository inquiries. CDC ID will also work.}

\item{org}{Name of CDC division, center, or office.}

\item{version}{Current release version of the software.}

\item{status}{Project development status (e.g., Maintained, Deprecated).}

\item{keywords}{Comma-separated list of technologies, domains, or topics.}

\item{labor_hours}{Estimated total person-hours for development.}

\item{contract}{Contract number associated with the project (if applicable).}

\item{exemption}{One of the five allowed values for restricted access.}

\item{exemption_justification}{Required explanation if an exemption is claimed. Use "ei" for EI group specific justification.}
}
\value{
An edit to the README.md file of the repository of the project you are in.
}
\description{
Add metadata marker based on SHARE IT Act of 2024 to be read by OCIO AI scanner. To be used as is and not in a script (e.g. in the console of an RStudio workstation).
}
\details{
\tabular{ll}{
   Exemption \tab Description \cr
   exemptByLaw \tab Code is exempt from sharing due to legal restrictions \cr
   exemptByNationalSecurity \tab Code is exempt due to national security concerns \cr
   exemptByAgencySystem \tab Code is exempt as it supports internal agency systems \cr
   exemptByMissionSystem \tab Code is exempt as it supports critical mission operations \cr
   exemptByCIO \tab Code has a special exemption approved by the CDC CIO \cr
}
}
\examples{
\dontrun{
add_metadata_marker("zzz0", exemption = "exemptByCIO", exemption_justification = "ei")
}

}
