% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_epc.R
\name{calculate_epc}
\alias{calculate_epc}
\title{Calculate Exposure Point Concentrations (EPCs)}
\usage{
calculate_epc(
  obs = NULL,
  cen = NULL,
  conf.level = 0.9,
  sigfig = 4,
  testForNormal = TRUE,
  useDefaultSeed = TRUE
)
}
\arguments{
\item{obs}{A numeric vector}

\item{cen}{A logical vector pertaining to censoring of obs. TRUE if obs is censored.}

\item{conf.level}{A numeric vector from 0 to 1 for confidence level. Default is 0.90 for computing 95UCL.}

\item{sigfig}{A numeric value for the number of significant figures for the outputs of the function. Default is 4.}

\item{testForNormal}{Logical. If you want to test for normal distribution of your obs. Default is TRUE}

\item{useDefaultSeed}{Logical. The Default TRUE uses custom seed similar to EPCTool, which is sum of all concentrations. For testing, either upload separate files of individual data subsets to EPCTool or sum all obs in clean dataset and set to seed.}
}
\value{
A data.frame class object.
\itemize{
\item \code{function_used}: the character string representing the type of calculation used for the EPC.
\item \code{mean}: numeric value of the estimated mean based on \code{function_used}.
\item \code{sd}: numeric value of the estimated standard deviation based on \code{function_used}.
\item \code{median}: numeric value of the estimated median based on \code{function_used}, if available.
\item \code{epc}: numeric value of the EPC based on \code{function_used}.
\item \code{mean_lci}: numeric value of the lower confidence interval based on \code{conf.level}, based on \code{function_used}.
\item \code{mean_uci}: numeric value of the upper confidence interval based on \code{conf.level}, based on \code{function_used}. Technically the EPC.
\item \code{notes}: character string of the underlying process that outputted the returned values.
\item \code{qcontrol}: character string of the potential errors related to the data in relation to the returned values.
\item \code{normal_dist}: numeric value of the Cox-value for normal distribution.
\item \code{lognorm_dist}: numeric value of the Cox-value for log-normal distribution.
\item \code{gamma_dist}: numeric value of the Cox-value for gamma distribution.
\item \code{best_dist}: either Normal, Gamma, or Lognormal, depending on what is chosen.
\item \code{dist_iqr}: character string of the IQR (interquartile range, i.e. 25th and 75th percentile) of the distribution, if avaliable.
\item \code{mean_ci}: character string of the confidence interval based on \code{conf.level}.
}
}
\description{
Calculate exposure point concentrations according to ATSDR Division of Community Health Investigations Exposure Point Concentration Guidance for Discrete Sampling (2023). Last updated 09/21/2023.
}
\examples{
results <- rexp(n = 15, rate = 1)
nondetects <- results<0.5
calculate_epc(obs = results, cen = nondetects)

data.frame(results, nondetects, group = rep(c("A", "B"), 8)) \%>\% group_by(group) \%>\% summarise(epc = calculate_epc(obs = results, cen = nondetects)$epc)


}
