% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_boxplot.R
\name{example_boxplot}
\alias{example_boxplot}
\title{Create Example Boxplot}
\usage{
example_boxplot(
  x = NULL,
  plot.color = "black",
  bracket.color = "darkgray",
  text.color = "black",
  title = "",
  caption = ""
)
}
\arguments{
\item{x}{An optional numerical vector if you want to import your own data instead}

\item{plot.color}{A character string denoting a color for the boxplot generated}

\item{bracket.color}{A character string denoting a color for the drawn lines generated}

\item{text.color}{A character string denoting a color for the text generated}

\item{title}{A character string for an optional title}

\item{caption}{A character string for an optional caption}
}
\value{
A ggplot class object
}
\description{
Creates a boxplot skeleton which shows all parts of the boxplot with labels naming each part, with options to format the plot as you would like -- useful when you need an example of a boxplot for your appendix.
}
\examples{
example_boxplot()

}
