% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_atsdr_footnote.R
\name{flextable_atsdr_footnote}
\alias{flextable_atsdr_footnote}
\title{Flextable Footnotes based on ATSDR Style Manual (2019)}
\usage{
flextable_atsdr_footnote(flextable, data)
}
\arguments{
\item{flextable}{A flextable object.}

\item{data}{A dataset with comment, i, j variables.}
}
\value{
A flextable class object.
}
\description{
Create flextable footnotes according to ATSDR Style Manual (2019). This is a wrapper for the atsdr_footnote_symbol() function.
}
\examples{
library(tidyverse)
library(flextable)

example_dataset <- iris |>
 flextable::flextable() |>
 flextable::set_caption(flextable::as_paragraph("Iris Title"))

add_footnotes <- data.frame(
 comment = c("column 3", "location (2,3)", "location (1,1)", "column 2", "title"),
 i = c(NA, 2, 1, NA, NA),
 j = c(3, 3, 1, 2, NA)
)

example_dataset |> flextable_atsdr_footnote(add_footnotes)
}
