% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_atsdr_footnote.R
\name{flextable_atsdr_footnote}
\alias{flextable_atsdr_footnote}
\title{Flextable Footnotes based on ATSDR Style Manual (2019)}
\usage{
flextable_atsdr_footnote(flextable, data)
}
\arguments{
\item{flextable}{A flextable object.}

\item{data}{A dataset with comment, i, j variables.}
}
\value{
A flextable class object.
}
\description{
Create flextable footnotes according to ATSDR Style Manual (2019). This is a wrapper for the atsdr_footnote_symbol() function.
}
\details{
The data argument needs a data.frame object that consists of three vectors labeled as comment, i, and j.

The i and j column should be a single number which represent the row (i) and column (j) of where the footnote is placed. If both are NA, then the footnote will be for the title, or caption in flextable terms. If the row (i) is NA but not the column (j), then the footnote will be for the column name, or header in flextable terms.

The comment column should be a character string and have the description for the footnote.
}
\examples{
library(tidyverse)
library(flextable)

example_dataset <- head(iris) |>
  flextable::flextable() |>
  flextable::set_caption(flextable::as_paragraph("Iris Title"))

add_footnotes <- data.frame(
  comment = c("column 3", "location (2,3)", "location (1,1)", "column 2", "title"),
  i = c(NA, 2, 1, NA, NA),
  j = c(3, 3, 1, 2, NA)
)

#example_dataset |> flextable_atsdr_footnote(add_footnotes)

}
